web/requirements.txt
django==5.1.1
fpdf2==2.7.9
Pillow==10.4.0
python-dateutil
gunicorn==23.0.0
icalevents==0.1.29
uharfbuzz==0.41.0
